feat(templates): add ConnectionTemplate
[Fixes #177610479]
