+++ src/index.tsx
@@ -7,6 +7,7 @@ export * from './link';
 export * from './markdown';
 export * from './pager';
 export * from './outline';
+export * from './relay';
 export * from './renderers';
 export * from './routing';
 export * from './simple-error-boundary';
