+++ src/templates/connection-template/connection-template.tsx
@@ -1,14 +1,17 @@
 import React from 'react';
 
-import {ConfigureFunction, Maybe} from '../..';
-import {Stringify} from '../../stringify';
+import {ConnectionLike, IdType, Maybe, NodeLike, useEdgeNodes} from '../..';
+import {
+  ConfigureFunction,
+  Configurer,
+  FieldConfigurationProvider,
+} from '../configuration';
+import {UnboundArrayTemplate} from '../array-template';
 import {
   FieldWrapper as FieldWrapperType,
   ItemWrapper as ItemWrapperType,
   Wrapper as WrapperType,
 } from '../support';
-
-import {ConnectionLike, NodeLike} from './types';
 
 export interface ConnectionTemplateProps<N extends NodeLike, PI> {
   connection: Maybe<ConnectionLike<N, PI>>;
@@ -18,6 +21,15 @@ export interface ConnectionTemplateProps<N extends NodeLike, PI> {
   FieldWrapper: FieldWrapperType<N>;
 }
 
-export const ConnectionTemplate = <N extends NodeLike, PI>(
-  props: ConnectionTemplateProps<N, PI>
-) => <Stringify>{props}</Stringify>;
+export const ConnectionTemplate = <N extends NodeLike, PI>({
+  connection,
+  ...rest
+}: ConnectionTemplateProps<N, PI>) => {
+  const nodes = useEdgeNodes(connection?.edges);
+
+  // @ts-expect-error - tsc can't tell that "id" is a key of N rather than just
+  // an arbitrary string. If I use NodeLike instead of N, it works.
+  const idField: IdType<N> = 'id';
+
+  return <ArrayTemplate idField={idField} data={nodes} {...rest} />;
+};
